Move the fetch command to the fetcher service.
